Canonicalize corpus WIP checkpoint + ipow pass
Checkpoint of in-progress work on the extended branch:

- Canonicalization npbench+polybench pipeline: statement-level passes
  (split_statements, perfect_loop_nesting), induction-variable and
  stride-permutation extensions, break-anti-dependence, buffer expansion,
  state-fusion/map-fusion reduction-ordering guards, LiftEinsum, finalize.
  Corpus ground truth: 44/54 canon bit-exact; xfail list reconciled to
  reality (contour_integral now passes; mandelbrot1/2 + stockham_fft are
  broken ports; cavity_flow + azimint_naive/nbody + correlation tracked
  as canon gaps).

- ipow: relax_integer_powers pass (Pow -> ipow for non-negative integer
  exponents) + symbolic ipow Function, wired into SIMPLIFY_PASSES; fixes
  stockham symbolic-exponent codegen. Matches the shipped relax_ipow branch.

- Vectorization corpus: multidim simplify+vectorize npbench/polybench
  suites, masked-write / multi-output tasklet normalization.

WIP: azimint_naive/nbody (in-nsdfg atomic reduction lost on
MapToForLoop sequentialization), cavity_flow, correlation still open.

+++ tests/corpus/npbench/structured_grids/cavity_flow.py
@@ -51,7 +51,7 @@ def pressure_poisson_np(nit, p, dx, dy, b):
 def reference(nx, ny, nt, nit, u, v, dt, dx, dy, p, rho, nu):
     un = np.empty_like(u)
     vn = np.empty_like(v)
-    b = np.zeros((ny, nx))
+    b = np.zeros((ny, nx), dtype=np.float32)
     for n in range(nt):
         un = u.copy()
         vn = v.copy()
@@ -105,7 +105,7 @@ def kernel(nt: dc.int64, u: dc_float[ny, nx], v: dc_float[ny, nx], dt: dc_float,
            p: dc_float[ny, nx], rho: dc_float, nu: dc_float):
     un = np.empty_like(u)
     vn = np.empty_like(v)
-    b = np.zeros((ny, nx))
+    b = np.zeros((ny, nx), dtype=np.float32)
     for n in range(nt):
         un[:] = u.copy()
         vn[:] = v.copy()
